Change implementation selection
Change the implementation selection: selection is now by calling package
instead of being with lexical scope ($Number::Phone::FR::Class).
This makes the selection more predictable, and is more
module-friendly.

Implementation selection is stored for each package and the calling
stack for ->new is examined to find the top package that selected a
particular implementation.